+++ iOS_SDK/OneSignalSDK/OneSignal_XCFramework/OneSignal.xcframework/ios-arm64_i386_x86_64-simulator/OneSignal.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module OneSignal {
+  umbrella header "OneSignal.h"
+
+  export *
+  module * { export * }
+}
