+++ iOS_SDK/OneSignalSDK/OneSignal_XCFramework/OneSignal.xcframework/ios-arm64_x86_64-maccatalyst/OneSignal.framework/Modules
@@ -1,0 +1,1 @@
+Versions/Current/Modules
