--- iOS_SDK/OneSignalSDK/OneSignal_XCFramework/OneSignal.xcframework/ios-i386_x86_64-simulator/OneSignal.framework/_CodeSignature/CodeResources
@@ -1,132 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>files</key>
-	<dict>
-		<key>Headers/OneSignal.h</key>
-		<data>
-		EVXphtkVA0pG/wCKXUqT8Br0olg=
-		</data>
-		<key>Info.plist</key>
-		<data>
-		HZ1/QrfvpUtCOkmBBcYC98xNQVs=
-		</data>
-		<key>Modules/module.modulemap</key>
-		<data>
-		fJj0ATF9ohjya70dUZjH6i9w44o=
-		</data>
-	</dict>
-	<key>files2</key>
-	<dict>
-		<key>Headers/OneSignal.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			EVXphtkVA0pG/wCKXUqT8Br0olg=
-			</data>
-			<key>hash2</key>
-			<data>
-			g07y/lXKM6H7TbypgKgFMT7K/hNdd1g4dcYnbX+edSE=
-			</data>
-		</dict>
-		<key>Modules/module.modulemap</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			fJj0ATF9ohjya70dUZjH6i9w44o=
-			</data>
-			<key>hash2</key>
-			<data>
-			Gn6ZaR3ERppbMM1cYIPVzhn7nRrRbOZoQYYdkOd4nJw=
-			</data>
-		</dict>
-	</dict>
-	<key>rules</key>
-	<dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Base\.lproj/</key>
-		<dict>
-			<key>weight</key>
-			<real>1010</real>
-		</dict>
-		<key>^version.plist$</key>
-		<true/>
-	</dict>
-	<key>rules2</key>
-	<dict>
-		<key>.*\.dSYM($|/)</key>
-		<dict>
-			<key>weight</key>
-			<real>11</real>
-		</dict>
-		<key>^(.*/)?\.DS_Store$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>2000</real>
-		</dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Base\.lproj/</key>
-		<dict>
-			<key>weight</key>
-			<real>1010</real>
-		</dict>
-		<key>^Info\.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^PkgInfo$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^embedded\.provisionprofile$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^version\.plist$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-	</dict>
-</dict>
-</plist>
