--- iOS_SDK/OneSignalSDK/OneSignal_XCFramework/OneSignal.xcframework/ios-x86_64-maccatalyst/OneSignal.framework/Versions/Current
@@ -1,1 +1,0 @@
-A
